feat(Demo): add TagOverTextOCR component
