+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/Demo/TagOverTextOCR.scss
@@ -1,0 +1,8 @@
+﻿.tag-over-text-ocr{
+  &__result{
+    word-break: break-word;
+  }
+  &__ocr-application{
+    text-align: center;
+  }
+}
